+++ modules/CompanyUser/Resources/routes/api.php
@@ -4,6 +4,7 @@ use Illuminate\Support\Facades\Route;
 use Modules\CompanyUser\Controllers\CompanyUserController;
 use Modules\CompanyUser\Controllers\CompanyUserProfileController;
 use Modules\User\Controllers\UserController;
+use Modules\RoleAndPermission\Enums\Permission;
 
 Route::group(['middleware' => ['auth:api',\Stancl\Tenancy\Middleware\InitializeTenancyByRequestData::class]], function () {
     Route::group(["prefix"=>"brokers"],function (){
@@ -23,7 +24,7 @@ Route::group(['middleware' => ['auth:api',\Stancl\Tenancy\Middleware\InitializeT
         Route::post('/', [\Modules\CompanyUser\Controllers\ClientController::class, 'store']);
 
     });
-    Route::get('/', [CompanyUserController::class, 'index']);
+    Route::get('/', [CompanyUserController::class, 'index'])->permission(Permission::USER_LIST());
     Route::get('/widgets', [CompanyUserController::class, 'widgets']);
     Route::get('/roles', [CompanyUserController::class, 'roles']);
     //Route::post('/export', [CompanyUserController::class, 'export'])->name('company-users.export');
@@ -32,13 +33,13 @@ Route::group(['middleware' => ['auth:api',\Stancl\Tenancy\Middleware\InitializeT
     Route::post('/validate-photo/{id?}', [CompanyUserProfileController::class, 'validatePhoto']);
     Route::post('/upload-photo/{id?}', [CompanyUserProfileController::class, 'uploadPhoto']);
     Route::put('/data-info/{id?}', [CompanyUserProfileController::class, 'updateDataInfo']);
-    Route::put('/contact-info/{id?}', [CompanyUserProfileController::class, 'updateContactInformation']);
-    Route::post('/identity-data/{id?}', [CompanyUserProfileController::class, 'identityData']);
+    Route::put('/contact-info/{id?}', [CompanyUserProfileController::class, 'updateContactInformation'])->permission(Permission::USER_PROFILE_CONTACT_UPDATE(),Permission::PROFILE_CONTACT_INFO_UPDATE(),Permission::PROFILE_ADDRESS_INFO_UPDATE());
+    Route::post('/identity-data/{id?}', [CompanyUserProfileController::class, 'identityData'])->permission(Permission::USER_PROFILE_IDENTITY_UPDATE(),Permission::PROFILE_PASSPORT_INFO_UPDATE(),Permission::PROFILE_BORDER_NUMBER_UPDATE(),Permission::PROFILE_RESIDENCE_INFO_UPDATE());
     Route::post('/send-otp/{id?}', [CompanyUserProfileController::class, 'sendOtp']);
     Route::post('/validate-otp/{id?}', [CompanyUserProfileController::class, 'validateOtp']);
     Route::get('/show-data-info/{id?}', [CompanyUserProfileController::class, 'showDataInfo']);
-    Route::get('/show-contact-information/{id?}', [CompanyUserProfileController::class, 'showContactInformation']);
-    Route::get('/show-identity-data/{id?}', [CompanyUserProfileController::class, 'showidentityData']);
+    Route::get('/show-contact-information/{id?}', [CompanyUserProfileController::class, 'showContactInformation'])->permission(Permission::USER_PROFILE_CONTACT_VIEW(),Permission::PROFILE_CONTACT_INFO_VIEW(),Permission::PROFILE_ADDRESS_INFO_VIEW());
+    Route::get('/show-identity-data/{id?}', [CompanyUserProfileController::class, 'showidentityData'])->permission(Permission::USER_PROFILE_IDENTITY_VIEW(),Permission::PROFILE_PASSPORT_INFO_VIEW(),Permission::PROFILE_BORDER_NUMBER_VIEW(),Permission::PROFILE_RESIDENCE_INFO_VIEW());
 
 
     Route::get('/widget/user/{id}', [CompanyUserProfileController::class, 'widget']);
@@ -46,16 +47,16 @@ Route::group(['middleware' => ['auth:api',\Stancl\Tenancy\Middleware\InitializeT
 
     Route::get('/show-by-email', [CompanyUserController::class, 'showByEmail']);
     Route::post('/change-time-zone/{id}', [CompanyUserController::class, 'changeTimeZone']);
-    Route::post('/', [CompanyUserController::class, 'store']);
+    Route::post('/', [CompanyUserController::class, 'store'])->permission(Permission::USER_CREATE());
     Route::post('/validations', [CompanyUserController::class, 'validation']);
     Route::post('/check-email', [CompanyUserController::class, 'checkEmail']);
 
-    Route::get('/{id}', [CompanyUserController::class, 'show']);
-    Route::put('/{id}', [CompanyUserController::class, 'update']);
+    Route::get('/{id}', [CompanyUserController::class, 'show'])->permission(Permission::USER_VIEW());
+    Route::put('/{id}', [CompanyUserController::class, 'update'])->permission(Permission::USER_UPDATE());
     Route::post('/{id}/assign-role', [CompanyUserController::class, 'assignRoleForCompanies']);
     Route::post('/{id}/assign-role-for-current-company', [CompanyUserController::class, 'assignRoleForCurrentCompany']);
-    Route::delete('/{id}', [CompanyUserController::class, 'delete']);
+    Route::delete('/{id}', [CompanyUserController::class, 'delete'])->permission(Permission::USER_DELETE());
     Route::delete('/{id}/specific-role', [CompanyUserController::class, 'deleteForSpecificRole']);
-    Route::post('/export', [UserController::class, 'export'])->middleware("permission:user.list")->name("users.export");
+    Route::post('/export', [UserController::class, 'export'])->permission(Permission::USER_EXPORT())->name("users.export");
 
 });
